Move RoadModsRemovalArgs to RemoveRoadMod.h
