+++ src/OpenLoco/src/Windows/Construction/OverheadTab.cpp
@@ -3,6 +3,7 @@
 #include "Drawing/SoftwareDrawingEngine.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Road/CreateRoadMod.h"
+#include "GameCommands/Road/RemoveRoadMod.h"
 #include "GameCommands/Track/CreateTrackMod.h"
 #include "GameCommands/Track/RemoveTrackMod.h"
 #include "Graphics/ImageIds.h"
